# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Pertanian API",
        "description": "RESTful API untuk Sistem Informasi Pertanian - Dokumentasi lengkap untuk endpoint CRUD Laporan dan Bantuan dengan autentikasi Laravel Sanctum",
        "contact": {
            "name": "API Support",
            "email": "admin@sistempertanian.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "API Base URL"
        }
    ],
    "paths": {
...
        "/api/v1/bantuan": {
            "get": {
                "tags": [
                    "Bantuan"
                ],
                "summary": "Get all bantuan (filtered by user role)",
                "operationId": "e5e35d7ca6c7011d7ee0eee91dc7512b",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "disetujui",
                                "ditolak"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bantuan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bantuan"
                ],
                "summary": "Create new bantuan request",
                "operationId": "185df3f08e37e14823a3219937d80167",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "jenis_bantuan",
                                    "jumlah",
                                    "alasan"
                                ],
                                "properties": {
                                    "jenis_bantuan": {
                                        "type": "string",
                                        "example": "Pupuk Organik"
                                    },
                                    "jumlah": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "alasan": {
                                        "type": "string",
                                        "example": "Untuk meningkatkan produksi padi"
                                    },
                                    "tanggal_permintaan": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-12"
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "example": "Sangat mendesak"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bantuan request created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Permintaan bantuan berhasil dibuat"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "403": {
                        "description": "Unverified account"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/bantuan/{id}": {
            "get": {
                "tags": [
                    "Bantuan"
                ],
                "summary": "Get bantuan detail",
                "operationId": "efb15c6c02c6f8ea575c5adf4a4a142d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bantuan detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bantuan not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Bantuan"
                ],
                "summary": "Update bantuan (only pending status)",
                "operationId": "05a647be65af656c5a7022dff4941cad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "jenis_bantuan": {
                                        "type": "string",
                                        "example": "Bibit Padi"
                                    },
                                    "jumlah": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150
                                    },
                                    "alasan": {
                                        "type": "string",
                                        "example": "Updated reason"
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "example": "Updated notes"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bantuan updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bantuan berhasil diperbarui"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized or already processed"
                    },
                    "404": {
                        "description": "Bantuan not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bantuan"
                ],
                "summary": "Delete bantuan (only pending status)",
                "operationId": "580fb4770ef0e66a985edcac57665486",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bantuan deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bantuan berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized or already processed"
                    },
                    "404": {
                        "description": "Bantuan not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/bantuan/{id}/approve": {
            "post": {
                "tags": [
                    "Bantuan"
                ],
                "summary": "Approve bantuan (Petugas/Admin only)",
                "operationId": "90fee27392167dd4d4565d325ddce5e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "catatan": {
                                        "type": "string",
                                        "example": "Disetujui"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bantuan approved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bantuan disetujui"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/bantuan/{id}/reject": {
            "post": {
                "tags": [
                    "Bantuan"
                ],
                "summary": "Reject bantuan (Petugas/Admin only)",
                "operationId": "20d3d2326c689b19059a35f12398ce7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "alasan"
                                ],
                                "properties": {
                                    "alasan": {
                                        "type": "string",
                                        "example": "Kuota habis"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bantuan rejected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bantuan ditolak"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token from login endpoint (Example: 1|abcdefghijklmnopqrstuvwxyz)",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}